Bugfix loan requests
Loan requests didn't work properly because ModelMapper, by default,
attempts to match all that is matchable. That lead to the "accountId"
field in the CreateLoanDTO being mapped to the "id" field of the Loan
class. This then causes Hibernate to attempt to SQL update the Loan
instead of inserting it, which causes an optimistic locking exception
(or similar).

Signed-off-by: David Djuretanovic <djues3@gmail.com>